refactor: update EmailService import paths and enhance UsersModule providers
