+++ src/resources/financial/reports/reports.module.ts
@@ -13,7 +13,6 @@ import { AdminImportReportsController } from './controllers/admin/admin-import-r
 import { ImportReportsProcessor } from './processors/import-reports.processor';
 import { UserReportsProcessor } from './processors/user-reports.processor';
 import { BaseReportProcessor } from './processors/base-report.processor';
-import { EmailService } from 'src/resources/email/email.service';
 import { AdminUserReportsController } from './controllers/admin/admin-user-reports.controller';
 import { UsersModule } from 'src/resources/users/users.module';
 import { UserFinancialReportsController } from './controllers/user-financial-reports.controller';
@@ -61,7 +60,6 @@ import { AdminUnlinkedReportService } from './services/admin/admin-unlinked-repo
     LoggerTxtService,
     ProgressService,
     S3Service,
-    EmailService,
     // Processors
     ImportReportsProcessor,
     UserReportsProcessor,
